coord: split arrangement_state mod into dedicated file
This cuts down on the size of coord.rs, which is still too large.
